// src/models/Cooler.tsx
export interface Cooler {
    id: number;
    manufacturer: string;
    model: string;
    price: number;
    rating: number;
    photos: string[]
    type: string;
    fanCount: number;
    fanSize: number;
    backlight: boolean;
    material: string;
    radiatorSize: string;
    compatibility: string;
}